+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Added LocationArgument to cloud-bukkit
  - Added ServerArgument to cloud-velocity
  - Added LockableCommandManager to cloud-core
+ - Added VelocityCommandPreprocessor to cloud-velocity
 
 ## [1.0.2] - 2020-10-18